requirements.txt
aiohappyeyeballs
aiohttp
altair
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
distro
docstring-parser
faiss-cpu==1.7.4
flask-cors
flask-talisman
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-generativeai
googleapis-common-protos
groq
grpcio
gunicorn
httpcore
httplib2
httpx
jinja2
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-groq
langchain-text-splitters
langsmith
numpy
orjson
pandas
pillow
protobuf
pydantic
pypdf
python-dateutil
python-dotenv
pyyaml
requests
sqlalchemy
tenacity
tqdm
typing-extensions
urllib3
werkzeug